Add files via upload
zeroMIPS1.c is the scaling function implementing the Modified Modified IPS (Modified MPS is due to Pretzel). Now it stores the matrix as float and not double. It saves some RAM and does not really affect anything else.
myComp1.c is the comparison function for sorting
utmvMul1.c multiplies a matrix stored in several arrays by a vector; both are float. This function takes the most time spent on scaling.
mainMIPS1.c is an example of how to use zeroMIPS1c to scale a matrix.