Add files via upload
The new version of balancing functions. Two variants - based on either SCALE or RU symmetry preserving algorithm.
The matrix is still stored as float but the arithmetic operations are now double.
The default tolerance is now 1.0e-8. Takes a bit more time but seems to work well (does not effect the percentage of rows which need to be removed).